Validation function for form openings and layer order for opened forms
The forms now have validations for opening, so they will not be allowed to open if the game situation is not appropriate for it, as unfortunately due to VS, forms behind others can still be interacted with. Another limitation is forms being closed using the OS and the data not being saved, which could not be fixed. Also, there is now a fixed layer order of forms so if forms behind others are clicked, they will not move to the top.